+++ lib/kennel/subclass_tracking.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 module Kennel
   module SubclassTracking
+    TRACKED_CLASSES = []
+
     def recursive_subclasses
       subclasses + subclasses.flat_map(&:recursive_subclasses)
     end
@@ -13,6 +15,7 @@ module Kennel
 
     def inherited(child)
       super
+      TRACKED_CLASSES << child
       subclasses << child
     end
   end
