+++ test/kennel/models/built/synthetic_test_test.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+require_relative "../../../test_helper"
+
+SingleCov.covered!
+
+describe Kennel::Models::Built::SyntheticTest do
+end
