+++ test/kennel/models/slo_test.rb
@@ -21,7 +21,6 @@ describe Kennel::Models::Slo do
   end
 
   let(:project) { TestProject.new }
-  let(:id_map) { Kennel::IdMap.new }
   let(:expected_basic_json) do
     {
       name: "Foo\u{1F512}",
@@ -73,41 +72,6 @@ describe Kennel::Models::Slo do
         slo(groups: -> { ["foo"] }).build!.as_json,
         expected_basic_json
       )
-    end
-  end
-
-  describe "#resolve_linked_tracking_ids!" do
-    it "ignores empty caused by ignore_default" do
-      slo = slo(monitor_ids: -> { nil }).build!
-      slo.resolve_linked_tracking_ids!(id_map, force: false)
-      refute slo.as_json[:monitor_ids]
-    end
-
-    it "does nothing for hardcoded ids" do
-      slo = slo(monitor_ids: -> { [123] }).build!
-      slo.resolve_linked_tracking_ids!(id_map, force: false)
-      slo.as_json[:monitor_ids].must_equal [123]
-    end
-
-    it "resolves relative ids" do
-      slo = slo(monitor_ids: -> { ["#{project.kennel_id}:mon"] }).build!
-      id_map.set("monitor", "#{project.kennel_id}:mon", 123)
-      slo.resolve_linked_tracking_ids!(id_map, force: false)
-      slo.as_json[:monitor_ids].must_equal [123]
-    end
-
-    it "does not resolve missing ids so they can resolve when monitor was created" do
-      slo = slo(monitor_ids: -> { ["#{project.kennel_id}:mon"] }).build!
-      id_map.set("monitor", "#{project.kennel_id}:mon", Kennel::IdMap::NEW)
-      slo.resolve_linked_tracking_ids!(id_map, force: false)
-      slo.as_json[:monitor_ids].must_equal ["test_project:mon"]
-    end
-
-    it "fails with typos" do
-      slo = slo(monitor_ids: -> { ["#{project.kennel_id}:mon"] }).build!
-      assert_raises Kennel::UnresolvableIdError do
-        slo.resolve_linked_tracking_ids!(id_map, force: false)
-      end
     end
   end
